pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "odds-tracker"
version = "0.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.118.0",
    "gitpython>=3.1.45",
    "nfl-data-py>=0.3.2",
    "nflreadpy>=0.1.3",
    "numpy>=2.3.3",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sqlmodel>=0.0.25",
    "uvicorn>=0.37.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "odds-tracker",
    "pytest>=8.4.2",
    "testcontainers>=4.13.1",
]

[tool.setuptools.packages.find]
where = ['src']
include = ['oddstracker*']


[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # Pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by formatter)
]

[tool.pylint.messages_control]
disable = [
    "C0114", # missing-module-docstring
    "C0115", # missing-class-docstring
    "C0116", # missing-function-docstring
    "R0903", # too-few-public-methods
    "R0913", # too-many-arguments
    "R0914", # too-many-locals
    "W1203", # logging-format-interpolation
    "W0718", # exception-escape
    "W0511", # TODOs
]

[tool.pylint.basic]
good-names = [
    "i",
    "j",
    "k",
    "ex",
    "Run",
    "_",
    "x",
    "y",
    "z",
    "id",
    "db",
    "pk",
    "ok",
]


[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["test"]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(name)s: %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
filterwarnings = [
    "ignore::DeprecationWarning:testcontainers.*"
]